refactoring, still ned to refactor ServerReceiver and Logic.
