+++ Communication/Shared/Data/Rank.cs
