+++ Communication/Shared/Data/Room.cs
