+++ Communication/Shared/PeriodicTask.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Communication.Shared
+{
+    public class PeriodicTask
+    {
+        public static void StartNew(int miliseconds, Action action)
+        {
+            Task.Factory.StartNew(() =>
+            {
+                while (true)
+                {
+                    action();
+                    Thread.Sleep(miliseconds);
+                }
+            });   
+        }
+    }
+}
