+++ servertcp/servertcp.csproj
@@ -73,9 +73,11 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Receiver.cs" />
-    <Compile Include="Scrypt.cs" />
-    <Compile Include="Server.cs" />
+    <Compile Include="Security\Scrypt.cs" />
+    <Compile Include="ServerManagment\ServerLogic.cs" />
+    <Compile Include="ServerManagment\Server.cs" />
+    <Compile Include="ServerManagment\ServerReceiver.cs" />
+    <Compile Include="ServerManagment\ServerReceiverEvents.cs" />
     <Compile Include="Sql\SqlHelper.cs" />
     <Compile Include="Sql\SqlUserCommands.cs" />
   </ItemGroup>
