+++ sharpdj/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:SharpDj"
-                ShutdownMode="OnLastWindowClose"
+             ShutdownMode="OnLastWindowClose"
              StartupUri="View/Views/SdjMainView.xaml">
     <Application.Resources>
         <ResourceDictionary>
@@ -10,12 +10,18 @@
                 <ResourceDictionary x:Name="ThemeDictionary">
                     <ResourceDictionary.MergedDictionaries>
                         <ResourceDictionary Source="Themes/GlobalStyles.xaml" />
-                        <ResourceDictionary Source="/Themes/Colors.xaml"/>
-                        <ResourceDictionary Source="/Themes/Buttons.xaml"/>
-                        <ResourceDictionary Source="/Themes/UserTabStyles.xaml"/>
+                        <ResourceDictionary Source="/Themes/Colors.xaml" />
+                        <ResourceDictionary Source="/Themes/Buttons.xaml" />
+                        <ResourceDictionary Source="/Themes/UserTabStyles.xaml" />
                     </ResourceDictionary.MergedDictionaries>
                 </ResourceDictionary>
             </ResourceDictionary.MergedDictionaries>
+            
+            <Style TargetType="{x:Type Window}">
+                <Setter Property="FontFamily"
+                        Value="Roboto Light" />
+            </Style>
         </ResourceDictionary>
     </Application.Resources>
+   
 </Application>
