+++ sharpdj/ViewModel/Model/RoomSquareModel.cs
@@ -1,11 +1,9 @@
-﻿using System;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading.Tasks;
 using Communication.Shared;
 using Hik.Communication.Scs.Communication.Messages;
 using Newtonsoft.Json;
 using SharpDj.Core;
-using SharpDj.Enums;
 using SharpDj.Enums.Menu;
 using SharpDj.Models.Client;
 using YoutubeExplode;
